refactor: rename Fragment to SFragment
